Services/data_ingestion/requirements.txt
pymupdf>=24.3.0
pytesseract>=0.3.10
pillow>=10.0.0
python-dotenv>=0.19.0
numpy>=1.24.0
tqdm>=4.0.0